+++ app/html/user/settings/account.html
@@ -7,7 +7,7 @@
 @def main():
 <div id="container">
   <h1>@{_('%(user)s\'s account', user=user.name)}</h1>
-  <form  method="POST" id="edit-user-form" data-refresh='true' class="pure-form pure-form-aligned">
+  <form  method="POST" id="edit-user-form" data-refresh='true'  action="@{url_for('do.edit_account')}" class="ajaxform pure-form pure-form-aligned">
     @{form.csrf_token()!!html}
     <fieldset>
       <div class="pure-control-group">
@@ -32,6 +32,8 @@
         <label for="oldpassword">@{form.oldpassword.label.text}</label>
         @{form.oldpassword(placeholder=form.oldpassword.label.text, required=True, autocomplete="off")!!html}
       </div>
+      <div class="alert div-message"></div>
+      <div class="alert div-error"></div>
       <button type="submit" class="pure-button pure-button-primary" id="edituser-btnsubmit" data-prog="@{_('Editing...')}" data-success="@{_('Saved!')}">@{_('Save')}</button>
     </fieldset>
   </form>
